Refactor CollectionsOverview and related components for improved structure and functionality
- Simplified CollectionsOverview component by extracting CanvasScene, CreateCollectionForm, and OverlayControls into separate files.
- Introduced useCollectionsState hook to manage collections state and logic.
- Added ChartPanel, CollectionSelector, TimeRangeToggle, and usePortfolioData for enhanced portfolio management features.
- Implemented responsive design and improved user experience with loading states and error handling.
- Updated styling for better visual consistency across components.
